safe moved config models and utils to internal folder
